+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to the PrettyPHP extension will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.1.5] - 2023-01-23
+
+### Added
+
+- Add PHP executable path setting.
+
+### Changed
+
+- Update PrettyPHP to v0.1.6.
+- Pass the active editor's indentation options to PrettyPHP.
+- Log output written to STDERR to the developer console even if no error occurs.
+- Allow code being formatted to use PHP's short open tag (`<?`).
+
+### Fixed
+
+- Fix PHP startup errors appearing in formatted code.
+
 ## [0.1.4] - 2023-01-20
 
 - Initial release